[Unittest] merge test_cp_async_in_if_then_else into test_tir_transform_inject_ptx_async_copy
This PR merge two related unittests into one.
